Pathway analysis by aligator

This function performs pathway analysis for every pathway at a particular MSigDB gene-set category using the ALIGATOR algorithm, and returns pathway-specific unadjusted p-value and adjusted p-value by permutation tests.

Usage

aligator(snpGeneP,setType=2, 
    Nsample=5000, Btimes=1000, pcut=0.05, seed=1)

Arguments

snpGeneP

It specifies a data frame containing columns of 'snp', 'gene_id' and 'p', which lists all analyzed SNPs, SNP-mapped genes and GWAS SNP association p-value

setType

'setType' is a integer or numeric value of gene-set category at [0,20]. It specifies the gene set category of MSigDB, from which every gene set is tested for enrichment. The setType takes default value of 2 ("C2: curated gene sets"). The complete list of gene set category can be obtained by 'getSetType' function.

Nsample

The number of random samples generated for calculating pathway-specific unadjusted p-value

Btimes

The number of bootstrap procedures for conducting permutation tests to get pathway adjusted p-value

pcut

A numerical value. It defines the cut-point p-value of SNP association for selecting 'significant' SNPs in pathway test.

seed

The seed number for ALIGATOR-based pathway test

Details

This pathway analysis is based on ALIGATOR algorithm. It takes a data frame of 'snpGeneP' that contains a character vector 'snp', an integer/numeric vector 'gene_id' of SNP-mapped genes (Entrez gene ID) and a numeric vector 'p' of GWAS SNP association p-value. The analysis uses a preselected p-value threshold ('pcut') to select significant SNPs for testing enrichment of the mapped genes. The pathway-specific unadjusted p-value is calculated from randomly generated 'Nsample' samples, and pathway adjusted p-value is obtained by 'Btimes' bootstrap-based permutations.

Value

The returned value is a data frame. It contains a integer vector 'pid' of pathway ID, a numeric vector 'p' of pathway-specific unadjusted p-value and a numeric vector 'adj_p' of pathway adjusted p-value by permutations.

References

Holmans, P., et al. (2009). "Gene ontology analysis of GWA study data sets provides insights into the biology of bipolar disorder." Am J Hum Genet 85(1): 13-24.

Author

H. Mei; L Li

Note

See also

Examples

# data(T2DGWAS)
# path0=aligator(snpGeneP,Nsample=100, Btimes=10)
# head(path0)