Uniform score calculation

Calculation of uniform score for gene effects

Usage

uscore(gm)

Arguments

gm

An integer or numeric vector of gene measure

Details

For a numeric or integer vector of measures for gene effects, the function calculates its uniform score.

Value

A numeric vector of uniform score

References

Author

H Mei; L Li

Note

See also

Examples

rn=rnorm(100); 
uscore(rn)
#>   [1] 0.115 0.585 0.015 0.455 0.745 0.885 0.065 0.315 0.335 0.285 0.205 0.755
#>  [13] 0.975 0.085 0.675 0.055 0.225 0.415 0.695 0.165 0.645 0.605 0.125 0.765
#>  [25] 0.955 0.395 0.155 0.445 0.175 0.095 0.795 0.545 0.575 0.935 0.505 0.375
#>  [37] 0.045 0.295 0.265 0.845 0.495 0.195 0.425 0.305 0.635 0.995 0.485 0.725
#>  [49] 0.515 0.035 0.785 0.345 0.355 0.465 0.475 0.705 0.025 0.985 0.245 0.555
#>  [61] 0.855 0.185 0.875 0.325 0.135 0.145 0.835 0.525 0.665 0.075 0.105 0.595
#>  [73] 0.925 0.565 0.915 0.685 0.735 0.385 0.535 0.405 0.965 0.895 0.775 0.715
#>  [85] 0.215 0.865 0.005 0.365 0.625 0.235 0.615 0.825 0.815 0.435 0.655 0.945
#>  [97] 0.255 0.805 0.905 0.275

##calculation of uniform score for differential expression p-value of T2D-GWES genes.
# data("T2DGWES")
# escore<-uscore(T2DExpression$p)
# head(escore)